fix(messages): make message ID unique per stream
Previously, message IDs were globally unique, which was not intended. The unique constraints on the schema have been changed, as well as the delete messages sproc. Since Snicket is so new, I don't think anyone has shipped any apps with it yet, so a clean setup isn't going to be a problem.